requirements.txt
streamlit
streamlit_chat
streamlit_extras
hugchat
Pillow
